+++ types/fast_warmup_settings.go
@@ -2,6 +2,6 @@ package divan_types
 
 type FastWarmupSettings struct {
 	FastWarmupEnabled  bool `json:"fastWarmupEnabled"`
-	MinItemsThreshold  uint `json:"minItemsThreshold"`
-	MinMemoryThreshold uint `json:"minMemoryThreshold"`
+	MinItemsThreshold  uint64 `json:"minItemsThreshold"`
+	MinMemoryThreshold uint64 `json:"minMemoryThreshold"`
 }
