+++ types/interesting_stats.go
@@ -1,19 +1,19 @@
 package divan_types
 
 type InterestingStats struct {
-	CmdGet                   uint `json:"cmd_get"`
-	CouchDocsActualDiskSize  uint `json:"couch_docs_actual_disk_size"`
-	CouchDocsDataSize        uint `json:"couch_docs_data_size"`
-	CouchSpatialDataSize     uint `json:"couch_spatial_data_size"`
-	CouchSpatialDiskSize     uint `json:"couch_spatial_disk_size"`
-	CouchViewsActualDiskSize uint `json:"couch_views_actual_disk_size"`
-	CouchViewsDataSize       uint `json:"couch_views_data_size"`
-	CurrItems                uint `json:"curr_items"`
-	CurrItemsTot             uint `json:"curr_items_tot"`
-	EpBgFetched              uint `json:"ep_bg_fetched"`
-	GetHits                  uint `json:"get_hits"`
-	MemUsed                  uint `json:"mem_used"`
-	Ops                      uint `json:"ops"`
-	VbActiveNumNonResident   uint `json:"vb_active_num_non_resident"`
-	VbReplicaCurrItems       uint `json:"vb_replica_curr_items"`
+	CmdGet                   uint64 `json:"cmd_get"`
+	CouchDocsActualDiskSize  uint64 `json:"couch_docs_actual_disk_size"`
+	CouchDocsDataSize        uint64 `json:"couch_docs_data_size"`
+	CouchSpatialDataSize     uint64 `json:"couch_spatial_data_size"`
+	CouchSpatialDiskSize     uint64 `json:"couch_spatial_disk_size"`
+	CouchViewsActualDiskSize uint64 `json:"couch_views_actual_disk_size"`
+	CouchViewsDataSize       uint64 `json:"couch_views_data_size"`
+	CurrItems                uint64 `json:"curr_items"`
+	CurrItemsTot             uint64 `json:"curr_items_tot"`
+	EpBgFetched              uint64 `json:"ep_bg_fetched"`
+	GetHits                  uint64 `json:"get_hits"`
+	MemUsed                  uint64 `json:"mem_used"`
+	Ops                      uint64 `json:"ops"`
+	VbActiveNumNonResident   uint64 `json:"vb_active_num_non_resident"`
+	VbReplicaCurrItems       uint64 `json:"vb_replica_curr_items"`
 }
