+++ types/ports.go
@@ -1,10 +1,10 @@
 package divan_types
 
 type Ports struct {
-	Proxy     uint `json:"proxy"`
-	HttpsCAPI uint `json:"httpsCAPI"`
-	HttpsMgmt uint `json:"httpsMgmt"`
-	Direct    uint `json:"direct"`
-	DistTCP   uint `json:"distTCP"`
-	DistTLS   uint `json:"distTLS"`
+	Proxy     uint64 `json:"proxy"`
+	HttpsCAPI uint64 `json:"httpsCAPI"`
+	HttpsMgmt uint64 `json:"httpsMgmt"`
+	Direct    uint64 `json:"direct"`
+	DistTCP   uint64 `json:"distTCP"`
+	DistTLS   uint64 `json:"distTLS"`
 }
